Total rebuild of aplication
