+++ PathOfHero/settings.py
@@ -12,6 +12,7 @@ https://docs.djangoproject.com/en/4.1/ref/settings/
 import os
 from pathlib import Path
 
+
 # Build paths inside the project like this: BASE_DIR / 'subdir'.
 BASE_DIR = Path(__file__).resolve().parent.parent
 
@@ -37,8 +38,10 @@ INSTALLED_APPS = [
     'django.contrib.sessions',
     'django.contrib.messages',
     'django.contrib.staticfiles',
-    'main',
-
+    'hero_app',
+    'user_app',
+    'monster_app',
+    'gameplay_app',
 ]
 
 MIDDLEWARE = [
@@ -64,6 +67,7 @@ TEMPLATES = [
                 'django.template.context_processors.request',
                 'django.contrib.auth.context_processors.auth',
                 'django.contrib.messages.context_processors.messages',
+                'hero_app.currency_context.bank'
             ],
         },
     },
@@ -78,7 +82,7 @@ WSGI_APPLICATION = 'PathOfHero.wsgi.application'
 DATABASES = {
     'default': {
         'ENGINE': 'django.db.backends.postgresql_psycopg2',
-        'NAME': 'pathofhero',
+        'NAME': 'pathofhero_db',
         'HOST': 'localhost',
         'PASSWORD': 'coderslab',
         'USER': 'postgres',
