+++ gameplay_app/apps.py
@@ -1,0 +1,6 @@
+from django.apps import AppConfig
+
+
+class GameplayAppConfig(AppConfig):
+    default_auto_field = 'django.db.models.BigAutoField'
+    name = 'gameplay_app'
