+++ hero_app/models.py
@@ -1,6 +1,6 @@
 from django.db import models
 from django.contrib.auth.models import User
-# Create your models here.
+
 HERO_RACE = (
     (1, "Human"),
     (2, "Elf"),
@@ -55,9 +55,16 @@ class WeaponHero(models.Model):
 class ArmorHero(models.Model):
     hero = models.ForeignKey(Hero, on_delete=models.CASCADE)
     bought_armors = models.ForeignKey(Armor, on_delete=models.CASCADE)
+    selected = models.BooleanField(default=False)
 
 
-class UserCurrency(models.Model):
-    user = models.ForeignKey(User, on_delete=models.CASCADE)
-    gold = models.IntegerField(default=0)
-    diamonds = models.IntegerField(default=0)
+# class Monster(models.Model):
+#     name = models.CharField(max_length=16)
+#     level = models.IntegerField(default=0)
+#     strength = models.IntegerField()
+#     dexterity = models.IntegerField()
+#     wisdom = models.IntegerField()
+#     endurance = models.IntegerField()
+#     health_points = models.IntegerField()
+#     max_health_points = models.IntegerField()
+#     difficult = models.IntegerField()
