+++ hero_app/views.py
@@ -1,106 +1,16 @@
+from audioop import reverse
+
 from django.shortcuts import render, redirect
 from django.views import View
-from django.http import HttpResponse
-from main.forms import LoginForm, UserCreateForm, CreateHeroForm, ArmorAddForm, WeaponAddForm
-from django.contrib.auth.models import User
-from django.contrib.auth import authenticate, login, logout
-from main.models import HERO_RACE
-from main.models import Hero, Armor, Weapon, UserCurrency, ArmorHero, WeaponHero
+from hero_app.forms import CreateHeroForm, ArmorAddForm, WeaponAddForm
+from hero_app.models import HERO_RACE
+from hero_app.models import Hero, Armor, Weapon, ArmorHero, WeaponHero
+from user_app.models import UserCurrency
 from django.urls import reverse_lazy
 from django.contrib.auth.mixins import LoginRequiredMixin
 
 
 # Create your views here.
-
-
-class DemoPageView(View):
-
-    def get(self, request):
-        return render(request, 'demo_page.html')
-
-
-class MainView(LoginRequiredMixin, View):
-    login_url = reverse_lazy('login')
-
-    def get(self, request):
-        currency = UserCurrency.objects.get(user=request.user.id)
-        return render(
-            request,
-            'main.html',
-            context={
-                'currency': currency
-            })
-
-
-class LoginView(View):
-    def get(self, request):
-        form = LoginForm()
-        return render(request, 'login.html', context={
-                'form': form
-            }
-        )
-
-    def post(self, request):
-        form = LoginForm(request.POST)
-
-        if form.is_valid():
-            data = form.cleaned_data
-
-            username = data.get('username')
-            password = data.get('password')
-
-            # authentication
-            user = authenticate(
-                username=username,
-                password=password
-            )
-
-            if user:
-                # log in
-                login(request, user)
-                return redirect('/hero_list/')
-            else:
-                return HttpResponse(f"Błąd uwierzytelnienia. Podano nieprawidłowe poświadczenia.")
-
-
-class LogoutView(View):
-    def get(self, request):
-        logout(request)
-        return render(
-            request,
-            'logout.html'
-        )
-
-
-class UserCreateView(View):
-
-    def get(self, request):
-        form = UserCreateForm()
-        return render(
-            request,
-            'create_user.html',
-            context={
-                'form': form,
-            }
-        )
-
-    def post(self, request):
-        form = UserCreateForm(request.POST)
-
-        if form.is_valid():
-            data = form.cleaned_data
-
-            user = User.objects.create_user(
-                username=data.get('login'),
-                password=data.get('password'),
-                first_name=data.get('first_name'),
-                last_name=data.get('last_name'),
-                email=data.get('email')
-            )
-            UserCurrency.objects.create(
-                user=user
-            )
-            return redirect('/login/')
 
 
 class HeroListView(LoginRequiredMixin, View):
@@ -110,15 +20,12 @@ class HeroListView(LoginRequiredMixin, View):
         hero_list = Hero.objects.filter(user=request.user).order_by('is_alive', '-level')
         hero_list_count = hero_list.count()
 
-        currency = UserCurrency.objects.get(user=request.user.id)
-
         return render(
             request,
             'hero_list.html',
             context={
                 'hero_list': hero_list,
                 'number': hero_list_count,
-                'currency': currency
             })
 
 
@@ -127,14 +34,12 @@ class CreateHeroView(LoginRequiredMixin, View):
 
     def get(self, request):
         form = CreateHeroForm()
-        currency = UserCurrency.objects.get(user=request.user.id)
 
         return render(
             request,
             'create_hero.html',
             context={
                 'form': form,
-                'currency': currency,
             })
 
     def post(self, request):
@@ -160,7 +65,7 @@ class CreateHeroView(LoginRequiredMixin, View):
 
             request.session['actual_hero'] = hero.id
 
-            return redirect('/hero_list/')
+            return redirect(reverse('hero_list'))
 
 
 class HeroDetailView(LoginRequiredMixin, View):
@@ -170,7 +75,6 @@ class HeroDetailView(LoginRequiredMixin, View):
 
         hero = Hero.objects.get(id=hero_id)
         hero_race = HERO_RACE[hero.race - 1][1]
-        currency = UserCurrency.objects.get(user=request.user.id)
 
         bought_armors = ArmorHero.objects.filter(hero_id=hero_id)
         bought_weapons = WeaponHero.objects.filter(hero_id=hero_id)
@@ -181,7 +85,6 @@ class HeroDetailView(LoginRequiredMixin, View):
             context={
                 'hero': hero,
                 'hero_race': hero_race,
-                'currency': currency,
                 'armors': bought_armors,
                 'weapons': bought_weapons,
             })
@@ -190,7 +93,7 @@ class HeroDetailView(LoginRequiredMixin, View):
 
         request.session['actual_hero'] = hero_id
 
-        return redirect('/main/')
+        return redirect(reverse('main_site'))
 
 
 class ArmoryListView(LoginRequiredMixin, View):
@@ -206,14 +109,12 @@ class ArmoryListView(LoginRequiredMixin, View):
         armors = Armor.objects.all()
         armors_number = armors.count()
 
-        currency = UserCurrency.objects.get(user=request.user.id)
         return render(
             request,
             'armor_list.html',
             context={
                 'armors': armors,
                 'armors_number': armors_number,
-                'currency': currency
             })
 
     def post(self, request):
@@ -221,7 +122,7 @@ class ArmoryListView(LoginRequiredMixin, View):
         hero_id = request.POST.get('hero')
         request.session['actual_hero'] = hero_id
 
-        return redirect("/armory/")
+        return redirect(reverse('armory'))
 
 
 class ArmorDetailView(LoginRequiredMixin, View):
@@ -235,14 +136,12 @@ class ArmorDetailView(LoginRequiredMixin, View):
             return render(request, 'list_no_hero.html', {'hero_list': hero_list})
 
         armor = Armor.objects.get(id=armor_id)
-        currency = UserCurrency.objects.get(user=request.user.id)
 
         return render(
             request,
             'armor_detail.html',
             context={
                 'armor': armor,
-                'currency': currency
             })
 
     def post(self, request, armor_id):
@@ -250,7 +149,7 @@ class ArmorDetailView(LoginRequiredMixin, View):
         hero_id = request.POST.get('hero')
         request.session['actual_hero'] = hero_id
 
-        return redirect(f"/armor_details/{armor_id}/")
+        return redirect(reverse('armor_detail',armor_id))
 
 
 class AddArmorView(LoginRequiredMixin, View):
@@ -259,14 +158,12 @@ class AddArmorView(LoginRequiredMixin, View):
     def get(self, request):
 
         form = ArmorAddForm()
-        currency = UserCurrency.objects.get(user=request.user.id)
 
         return render(
             request,
             'add_armor.html',
             context={
                 'form': form,
-                'currency': currency
             })
 
     def post(self, request):
@@ -286,7 +183,7 @@ class AddArmorView(LoginRequiredMixin, View):
                 diamonds=data.get('diamonds')
             )
 
-            return redirect('/armory/')
+            return redirect(reverse('armory'))
 
 
 class SmithListView(LoginRequiredMixin, View):
@@ -297,19 +194,16 @@ class SmithListView(LoginRequiredMixin, View):
         hero_list = Hero.objects.filter(user=request.user).order_by('is_alive', '-level')
 
         if request.session.get('actual_hero') is None:
-            currency = UserCurrency.objects.get(user=request.user.id)
 
             return render(
                 request,
                 'list_no_hero.html',
                 context={
                     'hero_list': hero_list,
-                    'currency': currency,
                 })
 
         weapons = Weapon.objects.all().order_by('price', 'attack_bonus')
         weapons_number = weapons.count()
-        currency = UserCurrency.objects.get(user=request.user.id)
 
         return render(
             request,
@@ -317,7 +211,6 @@ class SmithListView(LoginRequiredMixin, View):
             context={
                 'weapons': weapons,
                 'weapons_number': weapons_number,
-                'currency': currency
             })
 
     def post(self, request):
@@ -325,7 +218,7 @@ class SmithListView(LoginRequiredMixin, View):
         hero_id = request.POST.get('hero')
         request.session['actual_hero'] = hero_id
 
-        return redirect("/smith/")
+        return redirect(reverse('smith'))
 
 
 class WeaponDetailView(LoginRequiredMixin, View):
@@ -335,36 +228,31 @@ class WeaponDetailView(LoginRequiredMixin, View):
         hero_list = Hero.objects.filter(user=request.user).order_by('is_alive', '-level')
 
         if request.session.get('actual_hero') is None:
-            currency = UserCurrency.objects.get(user=request.user.id)
 
             return render(
                 request,
                 'list_no_hero.html',
                 context={
                     'hero_list': hero_list,
-                    'currency': currency,
                 })
 
         weapon = Weapon.objects.get(id=weapon_id)
-        currency = UserCurrency.objects.get(user=request.user.id)
 
         return render(request, 'weapon_detail.html', {
             'weapon': weapon,
-            'currency': currency
         })
 
     def post(self, request, weapon_id):
         hero_id = request.POST.get('hero')
         request.session['actual_hero'] = hero_id
 
-        return redirect(f"/weapon_details/{weapon_id}/")
+        return redirect(reverse('weapon_details',weapon_id))
 
 
 class AddWeaponView(LoginRequiredMixin, View):
     login_url = reverse_lazy('login')
 
     def get(self, request):
-        currency = UserCurrency.objects.get(user=request.user.id)
         form = WeaponAddForm()
 
         return render(
@@ -372,7 +260,6 @@ class AddWeaponView(LoginRequiredMixin, View):
             'add_weapon.html',
             context={
                 'form': form,
-                'currency': currency
             })
 
     def post(self, request):
@@ -391,7 +278,7 @@ class AddWeaponView(LoginRequiredMixin, View):
                 diamonds=data.get('diamonds')
             )
 
-            return redirect('/smith/')
+            return redirect(reverse('smith'))
 
 
 class BuyWeaponView(LoginRequiredMixin, View):
@@ -408,7 +295,6 @@ class BuyWeaponView(LoginRequiredMixin, View):
             request,
             'buy_weapon.html',
             context={
-                'currency': currency,
                 'weapon': weapon,
                 'gold_balance': gold_balance,
                 'diamond_balance': diamonds_balance
@@ -428,7 +314,7 @@ class BuyWeaponView(LoginRequiredMixin, View):
 
         weapon.hero.add(hero)
 
-        return redirect('/smith/')
+        return redirect(reverse('smith'))
 
 
 class BuyArmorView(LoginRequiredMixin, View):
@@ -465,5 +351,4 @@ class BuyArmorView(LoginRequiredMixin, View):
 
         armor.hero.add(hero)
 
-        return redirect('/armory/')
-
+        return redirect(reverse('armory'))
