+++ templates/armor_detail.html
@@ -11,7 +11,7 @@ Login
        <li>Attack bonus: {{ armor.attack_bonus }}</li>
        <li>Damage reduction: {{ armor.damage_reduction }}</li>
        <li>Price: {{ armor.price }}</li>
-       <li><a href="/buy_armor/{{ armor.id }}/">Buy</a></li>
+       <li><a href="{% url 'buy_armor' armor.id %}">Buy</a></li>
    </ul>
 </form>
 {% endblock content %}
