+++ templates/armor_list.html
@@ -14,7 +14,7 @@ Add armor
                     <li>Name: {{ armor.name }}</li>
                     <li>Armor description: {{ armor.description }}</li>
                     <li>Price: {{ armor.price }}
-                    <a href="/armor_details/{{ armor.id }}/" class="details-button">Details</a></li>
+                    <a href={% url 'armor_detail' armor.id %} class="details-button">Details</a></li>
                 </ul>
             {% endfor %}
         {% endif %}
