+++ templates/demo_page.html
@@ -9,7 +9,8 @@
         <header>
           <h1>Path Of Hero</h1>
         </header>
-        <a href="/hero_list/" class="play-button">Play</a>
+        <a href="{% url 'hero_list' %}" class="play-button">Play</a>
+
         <footer>
           <p>&copy; Medieval Adventure Game 2023</p>
         </footer>
