+++ templates/hero_detail.html
@@ -1,6 +1,6 @@
 {% extends 'base.html' %}
 {% block title %}
-Login
+Hero detail
 {% endblock title %}
 {% block content %}
 <form method="post">{% csrf_token %}
