+++ templates/hero_list.html
@@ -12,7 +12,7 @@ Login
              <a href="/create_hero/" class="create-hero-button">Create hero</a>
             {% for hero in hero_list %}
                 <li> Name: {{ hero.name }} Level: {{ hero.level }} {{ hero.health_points }}/{{ hero.max_health_points }}
-                <a href="/hero_details/{{ hero.id }}/" class="details-button">Details</a></li>
+                <a href={% url 'main_site' hero.id %} class="details-button">Details</a></li>
 {#                <input type="number" value='{{hero.id}}' name="heroes" hidden >#}
 {#                <input type="submit" value="Select as active">#}
             {% endfor %}
