+++ templates/main.html
@@ -1,6 +1,6 @@
 {% extends 'base.html' %}
 {% block title %}
-Main site
+main_site site
 {% endblock title %}
 {% block content %}
 <form method="post">{% csrf_token %}
